tidy up before upgrade
