--- UnitTests/EntityObjects/SimpleEntity.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-
-namespace EntityObjects
-{
-    public class SimpleEntity 
-    {
-        public string Name { get; set; }
-        public string Description { get; set; }
-        public int Age { get; set; }
-    }
-}
