+++ tests/DomainObjects/ComplexDomainObject.cs
