+++ tests/DomainObjects/SimpleDomainObject.cs
