+++ tests/MapperTests/SimpleDomainObjectTests.cs
